Setup CI.
This originally took a lot of time and commits, and it's still not
perfect. 13 commits compared to the 28 previously in the repository is
too many for me, so I'm squashing them.

Tags now automatically trigger a release, uploading the cpack zip file.
Commits get artifacts uploaded, the resulting zip is not quite the same
as the release zip (it's missing a folder) but it's Close Enough For
Now.